Use Json.Decode.Pipeline.resolve to compute Change
Instead of writing custom decoders, decode all used values and then
compute the resulting Change afterwards. This has the added benefit of
allowing access to all relevant pieces of data, and is a prerequisite
for writing a User generator, which will be dependant on three separate
pieces of the change object from the server.